py/plugins/dev-local-vectorstore/pyproject.toml
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#
# SPDX-License-Identifier: Apache-2.0

[project]
authors = [
  { name = "Google" },
  { name = "Yesudeep Mangalapilly", email = "yesudeep@google.com" },
  { name = "Elisa Shen", email = "mengqin@google.com" },
  { name = "Niraj Nepal", email = "nnepal@google.com" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Environment :: Web Environment",
  "Framework :: AsyncIO",
  "Framework :: Pydantic",
  "Framework :: Pydantic :: 2",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries",
  "Typing :: Typed",
  "License :: OSI Approved :: Apache Software License",
]
dependencies = [
  "aiofiles>=24.1.0",
  "genkit",
  "pytest-mock",
  "pytest-asyncio",
  "strenum>=0.4.15; python_version < '3.11'",
]
description = "Genkit Local Vector Store Plugin"
keywords = [
  "genkit",
  "ai",
  "llm",
  "machine-learning",
  "artificial-intelligence",
  "generative-ai",
  "vector-store",
  "embeddings",
  "local",
]
license = "Apache-2.0"
name = "genkit-plugin-dev-local-vectorstore"
readme = "README.md"
requires-python = ">=3.10"
version = "0.5.1"

[project.urls]
"Bug Tracker"   = "https://github.com/firebase/genkit/issues"
Changelog       = "https://github.com/firebase/genkit/blob/main/py/plugins/dev-local-vectorstore/CHANGELOG.md"
"Documentation" = "https://firebase.google.com/docs/genkit"
"Homepage"      = "https://github.com/firebase/genkit"
"Repository"    = "https://github.com/firebase/genkit/tree/main/py"

[build-system]
build-backend = "hatchling.build"
requires      = ["hatchling"]

[tool.hatch.build.targets.wheel]
only-include = ["src/genkit/plugins/dev_local_vectorstore"]
sources      = ["src"]
